// WhichPharmaPortal/WhichPharmaPortal/ClientApp/src/models/ProductV1.ts
import { ProductStatus } from './ProductStatus';
import { ShortageInfo } from './ShortageInfo';

export interface ProductV1 {
    countryName?: string;
    id: string;
    country: string;
    countryCode: string;
    scrapingOrigin: string;
    scrapingOriginId: number;
    scrapingOriginUrl?: string;
    lastUpdate: string;
    name: string;
    activeSubstances: string[];
    atc: string;
    drugForm?: string;
    administrationForm?: string;
    productCode?: string;
    manufacturer?: string;
    strength?: string;
    package?: string;
    maNumber?: string;
    maHolder?: string;
    status?: ProductStatus;
    spcFileUrl?: string;
    pilFileUrl?: string;
    labelFileUrl?: string;
    documentsUrls?: string[];
    retailPrice?: number;
    reimbursementPrice?: number;
    wholesalePrice?: number;
    factoryPrice?: number;
    exFactoryPrice?: number;
    retailPriceWithVat?: number;
    reimbursementPriceWithVat?: number;
    wholesalePriceWithVat?: number;
    factoryPriceWithVat?: number;
    exFactoryPriceWithVat?: number;
    currencyCode?: string;
    shortageInfo?: ShortageInfo;
    notes?: string;
    administrationFormCategories?: string[];
    drugFormCategories?: string[];
}